requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.54.0
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.5.0
aiofiles==23.2.1

